Blank App
Reset app to blank canvas
