+++ ClientApp/assets/adaptive-icon.png
